Need to tweak the closed fraction stat
Right now, the simulation has a variable closed fraction, that tweaks how much closed space there is. But that closed space is a baseline condition, we are really varying a treatment, the open area, which has a greater cost in this case. This leads to some counterintuitive figure interpretation later, or at least a bit more friction. This version does not have that fix, but I realized it should when plotting effect sizes and had a hard time figure out if there was a model problem, or a my brain problem.